Playwright_Framework/Day 2/Mini Project/requirements.txt
pytest-playwright
pytest-html
pytest